fix(formatters/eslint_d): fix cli arguments
Doing a search on [the changelog for
eslint_d](https://github.com/mantoni/eslint_d.js/blob/master/CHANGES.md)
it looks like no option called `--fix-to-stdout` ever existed, so I'm
not sure how/if this ever worked. Besides that, it appears like the
`--stdin-filename` argument is not actually needed and this works fine
without it.